Adding streaming to vehicle page
